+++ index.js
@@ -1,4 +1,4 @@
-const CalendarShell = require("./CalendarShell");
+const CalendarShell = require("./src/CalendarShell");
 
 const shell = new CalendarShell();
 shell.start();
