+++ src/CalendarAuthorizer.js
