+++ src/CalendarManager.js
